Trees
Moved the location stuff to its own file and added a heap interface
for working with the various frequencies around the map. This has
been used to add TREE features, with the different biomes supporting
a varying number of trees in each region. I've defined a region to be
a 64x64 pixel area.