River banks
RIVER is no longer used as a biome, instead, a flag is used along
with a flag for RiverBank. Another parallel function has been
introduced to add river bank features to the locations so that they
can be drawn later. The biome is left in the state that CalcBiome
sets it. This mainly works except for the edges near OCEAN tiles.
The method should be extended to handle not just the OCEAN and BEACH
merge, but also the tiles in which adjacent tiles are different
biomes.